+++ packages/atomic/src/components.d.ts
@@ -5,8 +5,8 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { CategoryFacet, CategoryFacetState, Engine, Facet, FacetState, LogLevel, Result, ResultTemplateCondition } from "@coveo/headless";
-import { Bindings, I18nState } from "./utils/initialization-utils";
+import { Engine, LogLevel, Result, ResultTemplateCondition } from "@coveo/headless";
+import { Bindings } from "./utils/initialization-utils";
 import { i18n } from "i18next";
 import { InitializationOptions } from "./components/atomic-search-interface/atomic-search-interface";
 export namespace Components {
@@ -59,32 +59,8 @@ export namespace Components {
     }
     interface AtomicNumericFacet {
         "facetId": string;
-        /**
-          * Specifies the index field whose values the facet should use
-         */
         "field": string;
-        /**
-          * Whether or not the index should automatically generate options for the facet
-         */
-        "generateAutomaticRanges": boolean;
-        /**
-          * The displayed label for the facet
-         */
         "label": string;
-    }
-    interface AtomicNumericFacetRange {
-        /**
-          * The ending value for the numeric range
-         */
-        "end": number;
-        /**
-          * Specifies whether or not the end value should be included in the range
-         */
-        "endInclusive": boolean;
-        /**
-          * The starting value for the numeric range
-         */
-        "start": number;
     }
     interface AtomicPager {
         /**
@@ -111,22 +87,15 @@ export namespace Components {
         "engine": Engine;
         "result": Result;
     }
+    interface AtomicResultExcerpt {
+    }
     interface AtomicResultLink {
     }
     interface AtomicResultList {
         /**
-          * Whether to automatically retrieve an additional page of results and append it to the current results when the user scrolls down to the bottom of element
+          * A list of fields to include in the query results, separated by commas.
          */
-        "enableInfiniteScroll": boolean;
         "fieldsToInclude": string;
-        /**
-          * Css class for the list wrapper
-         */
-        "listClass": string;
-        /**
-          * Css class for a list element
-         */
-        "listElementClass": string;
     }
     interface AtomicResultTemplate {
         "conditions": ResultTemplateCondition[];
@@ -134,7 +103,16 @@ export namespace Components {
         "getConditions": () => Promise<ResultTemplateCondition[]>;
         "getFields": () => Promise<string[]>;
     }
+    interface AtomicResultUri {
+    }
     interface AtomicResultValue {
+        /**
+          * Which highlight should the value be highlighted with
+         */
+        "shouldHighlightWith"?: string;
+        /**
+          * Which result value should the component render
+         */
         "value": string;
     }
     interface AtomicResultsPerPage {
@@ -143,9 +121,9 @@ export namespace Components {
          */
         "choicesDisplayed": string;
         /**
-          * Initial choice for the number of result per page. Should be part of the `choicesDisplayed` option.
+          * Initial choice for the number of result per page. Should be part of the `choicesDisplayed` option. By default, the first value of choices displayed.
          */
-        "initialChoice": number;
+        "initialChoice"?: number;
     }
     interface AtomicSearchBox {
         "_id": string;
@@ -198,21 +176,6 @@ export namespace Components {
           * String key value
          */
         "value": string;
-    }
-    interface BaseSearch {
-        "_id": string;
-        "hideSubmitButton": boolean;
-        "leadingSubmitButton": boolean;
-        "moreValuesAvailable": boolean;
-        "placeholder": string;
-        "strings": I18nState;
-        "suggestionValues": {value: string}[];
-        "value": string;
-    }
-    interface FacetSearch {
-        "_id": string;
-        "facet": Facet | CategoryFacet;
-        "facetState": FacetState | CategoryFacetState;
     }
 }
 declare global {
@@ -288,12 +251,6 @@ declare global {
         prototype: HTMLAtomicNumericFacetElement;
         new (): HTMLAtomicNumericFacetElement;
     };
-    interface HTMLAtomicNumericFacetRangeElement extends Components.AtomicNumericFacetRange, HTMLStencilElement {
-    }
-    var HTMLAtomicNumericFacetRangeElement: {
-        prototype: HTMLAtomicNumericFacetRangeElement;
-        new (): HTMLAtomicNumericFacetRangeElement;
-    };
     interface HTMLAtomicPagerElement extends Components.AtomicPager, HTMLStencilElement {
     }
     var HTMLAtomicPagerElement: {
@@ -324,6 +281,12 @@ declare global {
         prototype: HTMLAtomicResultElement;
         new (): HTMLAtomicResultElement;
     };
+    interface HTMLAtomicResultExcerptElement extends Components.AtomicResultExcerpt, HTMLStencilElement {
+    }
+    var HTMLAtomicResultExcerptElement: {
+        prototype: HTMLAtomicResultExcerptElement;
+        new (): HTMLAtomicResultExcerptElement;
+    };
     interface HTMLAtomicResultLinkElement extends Components.AtomicResultLink, HTMLStencilElement {
     }
     var HTMLAtomicResultLinkElement: {
@@ -341,6 +304,12 @@ declare global {
     var HTMLAtomicResultTemplateElement: {
         prototype: HTMLAtomicResultTemplateElement;
         new (): HTMLAtomicResultTemplateElement;
+    };
+    interface HTMLAtomicResultUriElement extends Components.AtomicResultUri, HTMLStencilElement {
+    }
+    var HTMLAtomicResultUriElement: {
+        prototype: HTMLAtomicResultUriElement;
+        new (): HTMLAtomicResultUriElement;
     };
     interface HTMLAtomicResultValueElement extends Components.AtomicResultValue, HTMLStencilElement {
     }
@@ -390,18 +359,6 @@ declare global {
         prototype: HTMLAtomicTextElement;
         new (): HTMLAtomicTextElement;
     };
-    interface HTMLBaseSearchElement extends Components.BaseSearch, HTMLStencilElement {
-    }
-    var HTMLBaseSearchElement: {
-        prototype: HTMLBaseSearchElement;
-        new (): HTMLBaseSearchElement;
-    };
-    interface HTMLFacetSearchElement extends Components.FacetSearch, HTMLStencilElement {
-    }
-    var HTMLFacetSearchElement: {
-        prototype: HTMLFacetSearchElement;
-        new (): HTMLFacetSearchElement;
-    };
     interface HTMLElementTagNameMap {
         "atomic-breadcrumb-manager": HTMLAtomicBreadcrumbManagerElement;
         "atomic-category-facet": HTMLAtomicCategoryFacetElement;
@@ -415,15 +372,16 @@ declare global {
         "atomic-history": HTMLAtomicHistoryElement;
         "atomic-no-results": HTMLAtomicNoResultsElement;
         "atomic-numeric-facet": HTMLAtomicNumericFacetElement;
-        "atomic-numeric-facet-range": HTMLAtomicNumericFacetRangeElement;
         "atomic-pager": HTMLAtomicPagerElement;
         "atomic-query-error": HTMLAtomicQueryErrorElement;
         "atomic-query-summary": HTMLAtomicQuerySummaryElement;
         "atomic-relevance-inspector": HTMLAtomicRelevanceInspectorElement;
         "atomic-result": HTMLAtomicResultElement;
+        "atomic-result-excerpt": HTMLAtomicResultExcerptElement;
         "atomic-result-link": HTMLAtomicResultLinkElement;
         "atomic-result-list": HTMLAtomicResultListElement;
         "atomic-result-template": HTMLAtomicResultTemplateElement;
+        "atomic-result-uri": HTMLAtomicResultUriElement;
         "atomic-result-value": HTMLAtomicResultValueElement;
         "atomic-results-per-page": HTMLAtomicResultsPerPageElement;
         "atomic-search-box": HTMLAtomicSearchBoxElement;
@@ -432,8 +390,6 @@ declare global {
         "atomic-sort-dropdown": HTMLAtomicSortDropdownElement;
         "atomic-tab": HTMLAtomicTabElement;
         "atomic-text": HTMLAtomicTextElement;
-        "base-search": HTMLBaseSearchElement;
-        "facet-search": HTMLFacetSearchElement;
     }
 }
 declare namespace LocalJSX {
@@ -485,32 +441,8 @@ declare namespace LocalJSX {
     }
     interface AtomicNumericFacet {
         "facetId"?: string;
-        /**
-          * Specifies the index field whose values the facet should use
-         */
         "field"?: string;
-        /**
-          * Whether or not the index should automatically generate options for the facet
-         */
-        "generateAutomaticRanges"?: boolean;
-        /**
-          * The displayed label for the facet
-         */
         "label"?: string;
-    }
-    interface AtomicNumericFacetRange {
-        /**
-          * The ending value for the numeric range
-         */
-        "end": number;
-        /**
-          * Specifies whether or not the end value should be included in the range
-         */
-        "endInclusive"?: boolean;
-        /**
-          * The starting value for the numeric range
-         */
-        "start": number;
     }
     interface AtomicPager {
         /**
@@ -537,29 +469,31 @@ declare namespace LocalJSX {
         "engine": Engine;
         "result": Result;
     }
+    interface AtomicResultExcerpt {
+    }
     interface AtomicResultLink {
     }
     interface AtomicResultList {
         /**
-          * Whether to automatically retrieve an additional page of results and append it to the current results when the user scrolls down to the bottom of element
+          * A list of fields to include in the query results, separated by commas.
          */
-        "enableInfiniteScroll"?: boolean;
         "fieldsToInclude"?: string;
-        /**
-          * Css class for the list wrapper
-         */
-        "listClass"?: string;
-        /**
-          * Css class for a list element
-         */
-        "listElementClass"?: string;
     }
     interface AtomicResultTemplate {
         "conditions"?: ResultTemplateCondition[];
         "fieldsToInclude"?: string;
     }
+    interface AtomicResultUri {
+    }
     interface AtomicResultValue {
-        "value"?: string;
+        /**
+          * Which highlight should the value be highlighted with
+         */
+        "shouldHighlightWith"?: string;
+        /**
+          * Which result value should the component render
+         */
+        "value": string;
     }
     interface AtomicResultsPerPage {
         /**
@@ -567,7 +501,7 @@ declare namespace LocalJSX {
          */
         "choicesDisplayed"?: string;
         /**
-          * Initial choice for the number of result per page. Should be part of the `choicesDisplayed` option.
+          * Initial choice for the number of result per page. Should be part of the `choicesDisplayed` option. By default, the first value of choices displayed.
          */
         "initialChoice"?: number;
     }
@@ -621,27 +555,6 @@ declare namespace LocalJSX {
          */
         "value": string;
     }
-    interface BaseSearch {
-        "_id": string;
-        "hideSubmitButton"?: boolean;
-        "leadingSubmitButton"?: boolean;
-        "moreValuesAvailable"?: boolean;
-        "onClear"?: (event: CustomEvent<void>) => void;
-        "onSearch"?: (event: CustomEvent<void>) => void;
-        "onSelectValue"?: (event: CustomEvent<number>) => void;
-        "onShowMoreResults"?: (event: CustomEvent<void>) => void;
-        "onTextChange"?: (event: CustomEvent<string>) => void;
-        "placeholder"?: string;
-        "strings": I18nState;
-        "suggestionValues": {value: string}[];
-        "value": string;
-    }
-    interface FacetSearch {
-        "_id"?: string;
-        "facet": Facet | CategoryFacet;
-        "facetState": FacetState | CategoryFacetState;
-        "onSelectValue"?: (event: CustomEvent<number>) => void;
-    }
     interface IntrinsicElements {
         "atomic-breadcrumb-manager": AtomicBreadcrumbManager;
         "atomic-category-facet": AtomicCategoryFacet;
@@ -655,15 +568,16 @@ declare namespace LocalJSX {
         "atomic-history": AtomicHistory;
         "atomic-no-results": AtomicNoResults;
         "atomic-numeric-facet": AtomicNumericFacet;
-        "atomic-numeric-facet-range": AtomicNumericFacetRange;
         "atomic-pager": AtomicPager;
         "atomic-query-error": AtomicQueryError;
         "atomic-query-summary": AtomicQuerySummary;
         "atomic-relevance-inspector": AtomicRelevanceInspector;
         "atomic-result": AtomicResult;
+        "atomic-result-excerpt": AtomicResultExcerpt;
         "atomic-result-link": AtomicResultLink;
         "atomic-result-list": AtomicResultList;
         "atomic-result-template": AtomicResultTemplate;
+        "atomic-result-uri": AtomicResultUri;
         "atomic-result-value": AtomicResultValue;
         "atomic-results-per-page": AtomicResultsPerPage;
         "atomic-search-box": AtomicSearchBox;
@@ -672,8 +586,6 @@ declare namespace LocalJSX {
         "atomic-sort-dropdown": AtomicSortDropdown;
         "atomic-tab": AtomicTab;
         "atomic-text": AtomicText;
-        "base-search": BaseSearch;
-        "facet-search": FacetSearch;
     }
 }
 export { LocalJSX as JSX };
@@ -692,15 +604,16 @@ declare module "@stencil/core" {
             "atomic-history": LocalJSX.AtomicHistory & JSXBase.HTMLAttributes<HTMLAtomicHistoryElement>;
             "atomic-no-results": LocalJSX.AtomicNoResults & JSXBase.HTMLAttributes<HTMLAtomicNoResultsElement>;
             "atomic-numeric-facet": LocalJSX.AtomicNumericFacet & JSXBase.HTMLAttributes<HTMLAtomicNumericFacetElement>;
-            "atomic-numeric-facet-range": LocalJSX.AtomicNumericFacetRange & JSXBase.HTMLAttributes<HTMLAtomicNumericFacetRangeElement>;
             "atomic-pager": LocalJSX.AtomicPager & JSXBase.HTMLAttributes<HTMLAtomicPagerElement>;
             "atomic-query-error": LocalJSX.AtomicQueryError & JSXBase.HTMLAttributes<HTMLAtomicQueryErrorElement>;
             "atomic-query-summary": LocalJSX.AtomicQuerySummary & JSXBase.HTMLAttributes<HTMLAtomicQuerySummaryElement>;
             "atomic-relevance-inspector": LocalJSX.AtomicRelevanceInspector & JSXBase.HTMLAttributes<HTMLAtomicRelevanceInspectorElement>;
             "atomic-result": LocalJSX.AtomicResult & JSXBase.HTMLAttributes<HTMLAtomicResultElement>;
+            "atomic-result-excerpt": LocalJSX.AtomicResultExcerpt & JSXBase.HTMLAttributes<HTMLAtomicResultExcerptElement>;
             "atomic-result-link": LocalJSX.AtomicResultLink & JSXBase.HTMLAttributes<HTMLAtomicResultLinkElement>;
             "atomic-result-list": LocalJSX.AtomicResultList & JSXBase.HTMLAttributes<HTMLAtomicResultListElement>;
             "atomic-result-template": LocalJSX.AtomicResultTemplate & JSXBase.HTMLAttributes<HTMLAtomicResultTemplateElement>;
+            "atomic-result-uri": LocalJSX.AtomicResultUri & JSXBase.HTMLAttributes<HTMLAtomicResultUriElement>;
             "atomic-result-value": LocalJSX.AtomicResultValue & JSXBase.HTMLAttributes<HTMLAtomicResultValueElement>;
             "atomic-results-per-page": LocalJSX.AtomicResultsPerPage & JSXBase.HTMLAttributes<HTMLAtomicResultsPerPageElement>;
             "atomic-search-box": LocalJSX.AtomicSearchBox & JSXBase.HTMLAttributes<HTMLAtomicSearchBoxElement>;
@@ -709,8 +622,6 @@ declare module "@stencil/core" {
             "atomic-sort-dropdown": LocalJSX.AtomicSortDropdown & JSXBase.HTMLAttributes<HTMLAtomicSortDropdownElement>;
             "atomic-tab": LocalJSX.AtomicTab & JSXBase.HTMLAttributes<HTMLAtomicTabElement>;
             "atomic-text": LocalJSX.AtomicText & JSXBase.HTMLAttributes<HTMLAtomicTextElement>;
-            "base-search": LocalJSX.BaseSearch & JSXBase.HTMLAttributes<HTMLBaseSearchElement>;
-            "facet-search": LocalJSX.FacetSearch & JSXBase.HTMLAttributes<HTMLFacetSearchElement>;
         }
     }
 }