+++ packages/headless/doc-parser/mocks/mock-obj-entity.ts
@@ -7,6 +7,8 @@ export function buildMockObjEntity(config: Partial<ObjEntity> = {}): ObjEntity {
     desc: '',
     isOptional: false,
     members: [],
+    isTypeExtracted: false,
+    typeName: '',
     ...config,
   };
 }