+++ packages/headless/doc-parser/src/function-resolver.test.ts
@@ -1,6 +1,7 @@
 import {buildMockApiFunction} from '../mocks/mock-api-function';
 import {buildMockApiInterface} from '../mocks/mock-api-interface';
 import {buildMockApiPropertySignature} from '../mocks/mock-api-property-signature';
+import {buildMockApiDocComment} from '../mocks/mock-api-doc-comment';
 import {buildMockEntity} from '../mocks/mock-entity';
 import {buildMockEntryPoint} from '../mocks/mock-entry-point';
 import {
@@ -14,8 +15,13 @@ import {resolveFunction} from './function-resolver';
 describe('#resolveFunction', () => {
   it('resolves function with a an interface return type', () => {
     const entry = buildMockEntryPoint();
+    const comment = buildMockApiDocComment(
+      '/**\n * Creates a `NumericRangeRequest`.\n *\n * @param config - The options with which to create a `NumericRangeRequest`.\n *\n * @returns A new `NumericRangeRequest`.\n */\n'
+    );
+
     const fn = buildMockApiFunction({
       name: 'buildNumericRange',
+      docComment: comment,
       excerptTokens: [
         buildContentExcerptToken('declare function buildNumericRange(config: '),
         buildReferenceExcerptToken(
@@ -72,17 +78,21 @@ describe('#resolveFunction', () => {
     const optionsParam = buildMockObjEntity({
       name: 'config',
       type: 'NumericRangeOptions',
+      typeName: 'NumericRangeOptions',
+      desc: 'The options with which to create a `NumericRangeRequest`.',
       members: [startProperty],
     });
 
     const returnType = buildMockObjEntity({
       name: 'NumericRangeRequest',
       type: 'NumericRangeRequest',
+      typeName: 'NumericRangeRequest',
       members: [startProperty],
     });
 
     const funcEntity = buildMockFuncEntity({
       name: 'buildNumericRange',
+      desc: 'Creates a `NumericRangeRequest`.',
       params: [optionsParam],
       returnType,
     });