+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0-alpha.11](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.10...v0.7.0-alpha.11) (2021-02-21)
+
+
+### Bug Fixes
+
+* **headless:** support multiple author for a document for analytics ([#515](https://github.com/coveo/ui-kit/issues/515)) ([a0c7c48](https://github.com/coveo/ui-kit/commit/a0c7c486c2ee1d0e750579ec965ac9c606a31294))
+
+
+### Features
+
+* **atomic,headless:** add highlights, atomic-result-excerpt & atomic-result-uri components ([#524](https://github.com/coveo/ui-kit/issues/524)) ([2b73da1](https://github.com/coveo/ui-kit/commit/2b73da10bbfdf52d9ada6533b532b9ae7463cd98))
+
+
+
+
+
+# [0.7.0-alpha.10](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.9...v0.7.0-alpha.10) (2021-02-19)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [0.7.0-alpha.3](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.2...v0.7.0-alpha.3) (2021-02-17)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [0.7.0-alpha.2](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.1...v0.7.0-alpha.2) (2021-02-17)
+
+
+### Features
+
+* **headless:** added `ResultLinkAnalytics` ([#492](https://github.com/coveo/ui-kit/issues/492)) ([de45056](https://github.com/coveo/ui-kit/commit/de45056386fd0b487c835090bb68f37aab8652aa))
+
+
+
+
+
+# [0.7.0-alpha.1](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.0...v0.7.0-alpha.1) (2021-02-17)
+
+
+### Bug Fixes
+
+* **headless:** remove history initial state deconstruction ([#514](https://github.com/coveo/ui-kit/issues/514)) ([3e1035c](https://github.com/coveo/ui-kit/commit/3e1035cda793106f52ff4e48c2903eff5c55a733))
+
+
+
+
+
 # [0.7.0-alpha.0](https://github.com/coveo/ui-kit/compare/v0.6.1-alpha.0...v0.7.0-alpha.0) (2021-02-17)