+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0-alpha.11](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.10...v0.7.0-alpha.11) (2021-02-21)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.7.0-alpha.10](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.9...v0.7.0-alpha.10) (2021-02-19)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.7.0-alpha.9](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.8...v0.7.0-alpha.9) (2021-02-19)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.7.0-alpha.7](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.6...v0.7.0-alpha.7) (2021-02-18)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.7.0-alpha.4](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.3...v0.7.0-alpha.4) (2021-02-17)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.7.0-alpha.3](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.2...v0.7.0-alpha.3) (2021-02-17)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.7.0-alpha.2](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.1...v0.7.0-alpha.2) (2021-02-17)
+
+
+### Features
+
+* **headless:** added `ResultLinkAnalytics` ([#492](https://github.com/coveo/ui-kit/issues/492)) ([de45056](https://github.com/coveo/ui-kit/commit/de45056386fd0b487c835090bb68f37aab8652aa))
+
+
+
+
+
+# [0.7.0-alpha.1](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.0...v0.7.0-alpha.1) (2021-02-17)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 # [0.7.0-alpha.0](https://github.com/coveo/ui-kit/compare/v0.6.1-alpha.0...v0.7.0-alpha.0) (2021-02-17)
 
 **Note:** Version bump only for package headless-react